// src/gen-backend/types/activity-controller-feed.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ActivityControllerFeedQueryParams = {
    /**
     * @type string
    */
    accountId: string;
    /**
     * @type string
    */
    cursor: string;
    /**
     * @type string
    */
    limit: string;
};

export type ActivityControllerFeed200 = {
    /**
     * @type array
    */
    items: object[];
    /**
     * @type string
    */
    nextCursor: string | null;
};

export type ActivityControllerFeedQueryResponse = ActivityControllerFeed200;

export type ActivityControllerFeedQuery = {
    Response: ActivityControllerFeed200;
    QueryParams: ActivityControllerFeedQueryParams;
    Errors: any;
};